+++ .vscode/settings.json
@@ -1,0 +1,11 @@
+{
+    "protoc": {
+        "compile_all_path": "${workspaceRoot}/src",
+        "compile_on_save": true,
+        "options": [
+            "--proto_path=${workspaceRoot}/src",
+            "--python_out=src",
+            "--pyi_out=src"
+        ],
+    }
+}
