--- .vscode/tasks.json
@@ -1,13 +1,0 @@
-{
-    // See https://go.microsoft.com/fwlink/?LinkId=733558
-    // for the documentation about the tasks.json format
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "label": "regenerate proto",
-            "type": "shell",
-            "command": "protoc --python_out=src/fr24/proto --pyi_out=src/fr24/proto --proto_path src/fr24/proto request.proto",
-            "problemMatcher": []
-        }
-    ]
-}
