+++ src/fr24/proto/_nearest_flights.proto
@@ -1,0 +1,23 @@
+syntax = "proto3";
+
+import "fr24/proto/_common.proto";
+
+message Geolocation {
+    float lat = 1;
+    float lon = 2;
+}
+
+message NearestFlightsRequest {
+    Geolocation location = 1;
+    uint32 radius = 2;
+    uint32 limit = 3;
+}
+
+message NearbyFlight {
+    _common.Flight flight = 1;
+    uint32 distance = 2;
+}
+
+message NearestFlightsResponse {
+    repeated NearbyFlight flights_list = 1;
+}
