+++ src/fr24/static/bbox.py
